CMake and CLion stuff
