+++ CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.19)
+project(quadtree_amogufier)
+
+set(CMAKE_CXX_STANDARD 17)
+
+add_executable(quadtree_amogufier src/main.cpp src/Image.cpp)
