+++ src/lib/stb_image.h
