+++ src/lib/stb_image_write.h
